feat(api): backend refactoring
* Update folders structure - move apps to `modules` dir.

* Split library app to smaller components.

* Extract business logic to separated service layer.
