+++ api-backend/modules/catalog/consts.py
@@ -1,0 +1,1 @@
+LECTION_PREVIEW_SYMBOLS_COUNT = 255
