+++ api-backend/modules/core/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class LibraryConfig(AppConfig):
+class CoreConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "library"
+    name = "modules.core"
