+++ api-backend/modules/core/exports.py
@@ -1,0 +1,2 @@
+from .models import Lection, Country
+from .serializers import LectionSerializer, LectionReadSerializer
