+++ api-backend/sahajalib/urls.py
@@ -19,8 +19,9 @@ from django.urls import path, include
 def trigger_error(request):
     division_by_zero = 1 / 0
 
+
 urlpatterns = [
-    path('sentry-debug/', trigger_error),
-    # ...
+    path("admin/", admin.site.urls),
+    path("api/", include("modules.search.urls")),
+    path("api/", include("modules.catalog.urls"))
 ]
-urlpatterns += [path("admin/", admin.site.urls), path("api/", include("library.urls"))]
